🎭 feat: menu-bar dashboard with tool/model views + themes
Rework the macOS app into a glanceable menu-bar dropdown (MenuBarExtra,
LSUIElement so no Dock icon) — the user mainly wants to peek at usage, not
open a full app. Ports Tokei's design system (dark cards, gradient stroke,
materials, SF Symbols, sliding tabs, rings) as the visual foundation.

- Group-by dimension toggle: aggregate cards by tool OR by model; the
  expandable breakdown shows the other dimension.
- Distribution bar chart (Swift Charts) + per-card metrics (in/out/cache
  read+write/requests) + cache-hit ring + rate-limit bars when present.
- Time range switcher (today/yesterday/week/last week/month/year).
- Two themes — warm (coding energy) and cool (Codex-like calm) — toggled
  from the header and persisted via @AppStorage.
- Client-side aggregation from /stats/summary + /ratelimits; no cost shown.
- Fix: RateLimitDto optional counts use i32 so typeshare keeps optionality.

+++ crates/token9-contracts/src/lib.rs
@@ -75,15 +75,13 @@ pub struct RateLimitDto {
     pub provider: String,
     #[typeshare(serialized_as = "I54")]
     pub updated_at: i64,
-    #[typeshare(serialized_as = "I54")]
-    pub requests_limit: Option<i64>,
-    #[typeshare(serialized_as = "I54")]
-    pub requests_remaining: Option<i64>,
+    // Optional counts kept as i32 so typeshare preserves optionality
+    // (serialized_as would drop it); vendor rate-limit counts fit in i32.
+    pub requests_limit: Option<i32>,
+    pub requests_remaining: Option<i32>,
     pub requests_reset: Option<String>,
-    #[typeshare(serialized_as = "I54")]
-    pub tokens_limit: Option<i64>,
-    #[typeshare(serialized_as = "I54")]
-    pub tokens_remaining: Option<i64>,
+    pub tokens_limit: Option<i32>,
+    pub tokens_remaining: Option<i32>,
     pub tokens_reset: Option<String>,
 }
